services/medfiles/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.95.0",
    "httpx>=0.24.1",
    "minio>=7.1.14",
    "numpy>=1.24.3",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation>=0.60b1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-logging>=0.60b1",
    "opentelemetry-instrumentation-requests>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "pillow>=9.5.0",
    "pydantic>=2.0.0",
    "pydicom>=2.3.1",
    "pyjwt>=2.7.0",
    "pymongo>=4.3.3",
    "python-dotenv>=1.0.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.6",
    "redis>=4.5.4",
    "uvicorn>=0.22.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]

[tool.ruff]
target-version = "py312"
line-length = 200

[tool.ruff.lint]
select = [
    "E",   
    "W",   
    "F",   
    "I",   
    "UP",  
    "B",   
    "S",    
    "C4",  
    "DTZ",  
    "T20",  
    "ASYNC",
    "BLE",  
    "PL",   
    "RUF", 
]

ignore = [
    "S101",
    "PLR2004",
]

[tool.ruff.lint.isort]
known-first-party = ["medrax"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.pylint]
max-args = 8

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false